@@ -1,7 +1,7 @@
-import React from "react"
-import ReactDOM from "react-dom"
+import { StrictMode } from "react"
+import { createRoot } from "react-dom/client"
 import App from "app"
-import * as serviceWorker from "./serviceWorker"
+import reportWebVitals from "./reportWebVitals"
 import { createGlobalStyle } from "styled-components"
 
 const styled = { createGlobalStyle }
@@ -23,15 +23,16 @@ const GlobalStyle = styled.createGlobalStyle`
   }
 `
 
-ReactDOM.render(
-  <React.StrictMode>
+const container = document.getElementById("root")
+const root = createRoot(container)
+root.render(
+  <StrictMode>
     <App />
     <GlobalStyle />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </StrictMode>
 )
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister()
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals()
